feat(mcp): cut over to the SDK envelope and declare the two missing conformance fields
The SDK envelope has served all production traffic since 2026-08-06 12:15 UTC,
with parity verified live across 12 protocol cases and 95/98 zero-argument
tools. The flag has done its job. Keeping it means keeping a second envelope
nobody exercises, which is how the untested path becomes the one that breaks,
so mcpSdkEnvelopeEnabled and its branch are removed.

The malformed-input branch stays, and is not a rollback. The SDK's transport
answers 400 -32700 Parse error for the whole request, which is wrong three
ways: -32700 is reserved for JSON that did not parse (-32600 is Invalid
Request), a JSON-RPC error belongs inside a 200, and for a mixed batch the SDK
drops the valid members outright. That last one is data loss, so malformed
input is answered by dispatchMessage directly and always will be.

Two MCP 2025-11-25 fields were never emitted:

  Tool.execution.taskSupport  declared "forbidden" on every tool -- this
                              server registers no task store, and absent the
                              declaration a client learns that by attempting a
                              task-augmented call and failing.
  Implementation.websiteUrl   both on serverInfo, both what a client renders in
  Implementation.icons        a server picker. Icons point at real published
                              assets on our own origin.

Icons are server-level only. The spec allows them per tool, but all 224 would
carry the identical set -- roughly 45 KB added to a tools/list every client
holds in context, for nothing the handshake did not already say. Same
reasoning that refused tools/list pagination in #9648, and asserted so the
decision is explicit rather than an omission someone later "fixes".

Adding websiteUrl turned up three copies of the site origin (src/feeds.ts,
src/subnet-news.ts, and the icons wanted a fourth); collapsed into SITE_ORIGIN
in src/contracts.ts beside PRIMARY_DOMAIN.

The parity harness becomes a contract test: with one implementation left there
is nothing to diff against, so every expectation is now the literal response a
caller receives. That is the stronger form regardless -- a comparison passes
happily when both sides regress together, which initialize was caught doing
during the migration.

Closes #9680
Closes #9647
